Add Basic Authorization to EmployeeRst

+++ Order/src/main/java/com/helman/Webservice/EmployeeRst.java
@@ -5,7 +5,7 @@ package com.helman.Webservice;
 //@Date 2/28/21
 //@Time 3:13AM
 //        Created by Intellije IDEA
-//        Description:JPA-Criteria
+//        Description: With Basic Authentication
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -16,6 +16,8 @@ import com.helman.Dao.Employeedao;
 import com.helman.Entity.Employee;
 import com.helman.General.Logback;
 import javax.ws.rs.*;
+import javax.ws.rs.core.Context;
+import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import java.util.List;
@@ -23,17 +25,22 @@ import java.util.List;
 @Path("/employee")
 public class EmployeeRst {
     Employeedao employeedao = new Employeedao();
-    //http:localhost:8080/order/rest/employee/all
+    Security sec=new Security();
+    //http://localhost:8080/order/rest/employee/all
     @GET
     @Path("/all")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response findall(){
+    public Response findall(@Context HttpHeaders headers){
         List<Employee> employeeList = employeedao.findall();
+        Employee employee = new Employee();
         //return employeeList;
+        String encodUsrPwd = headers.getRequestHeader("Authorization").get(0).replaceFirst("Basic ", "");
+        if (!sec.basicAuthCheck(encodUsrPwd)) return Response.status(Response.Status.UNAUTHORIZED).entity("User or password is wrong").build();
+
         try{
             //filter some attributes
-            FilterProvider filters = new SimpleFilterProvider().addFilter("Employeefilter", SimpleBeanPropertyFilter.filterOutAllExcept
-                    ("employeeNumber", "lastName", "firstName", "email"));
+            FilterProvider filters = new SimpleFilterProvider().addFilter("Employeefilter",
+                    SimpleBeanPropertyFilter.filterOutAllExcept(employee.getfilters()));
             //(list)Obj -> Json
             //json = mapper(obj)
             String employeeJson = (new ObjectMapper()).writer(filters).withDefaultPrettyPrinter().writeValueAsString(employeeList);
@@ -43,18 +50,21 @@ public class EmployeeRst {
         }catch (JsonProcessingException e){
             Logback.logger.error("{}.{}|Exception:{}", this.getClass().getSimpleName(), Thread.currentThread().getStackTrace()[1].getMethodName(), e.getMessage());
             e.printStackTrace();
-            return Response.status(Response.Status.NOT_FOUND).entity(e.getMessage()).build();
+            return Response.status(Response.Status.UNAUTHORIZED).entity(e.getMessage()).build();
         }
     }
-    //http://localhost8080/order/rest/employee/find/1004
+    //http://localhost:8080/order/rest/employee/find/1004
     @GET
     @Path("/find/{employeeNumber}")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response findById(@PathParam("employeeNumber") Long empnum){
+    public Response findById(@PathParam("employeeNumber") Long empnum, @Context HttpHeaders headers){
         Employee employee = employeedao.findbyid(empnum);
+        String encodUsrPwd = headers.getRequestHeader("Authorization").get(0).replaceFirst("Basic ", "");
+        if (!sec.basicAuthCheck(encodUsrPwd)) return Response.status(Response.Status.UNAUTHORIZED).entity("User or password is wrong").build();
+
         try{
             FilterProvider filters = new SimpleFilterProvider().addFilter("Employeefilter",
-                    SimpleBeanPropertyFilter.filterOutAllExcept("employeeNumber", "lastName", "firstName", "email"));
+                    SimpleBeanPropertyFilter.filterOutAllExcept(employee.getfilters()));
             String employeeJson = (new ObjectMapper()).writer(filters).withDefaultPrettyPrinter().writeValueAsString(employee);
             Logback.logger.info("{}.{}|Try:Send record to RESTful", this.getClass().getSimpleName(), Thread.currentThread().getStackTrace()[1].getMethodName());
             return Response.status(Response.Status.OK).entity(employeeJson).build();
@@ -96,7 +106,10 @@ public class EmployeeRst {
     @Path("/update")
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
-    public Response update(Employee employee){
+    public Response update(Employee employee, @Context HttpHeaders headers){
+        String encodUsrPwd = headers.getRequestHeader("Authorization").get(0).replaceFirst("Basic ", "");
+        if (!sec.basicAuthCheck(encodUsrPwd)) return Response.status(Response.Status.UNAUTHORIZED).entity("User or password is wrong").build();
+
         try{
             Employee updatedEmployee = employeedao.findbyid(employee.getEmployeeNumber());
             updatedEmployee.setEmployeeNumber(employee.getEmployeeNumber());
@@ -119,7 +132,10 @@ public class EmployeeRst {
     @DELETE
     @Path("/{employeeNumber}")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response delete(@PathParam("employeeNumber") Long empnum){
+    public Response delete(@PathParam("employeeNumber") Long empnum, @Context HttpHeaders headers){
+        String encodUsrPwd = headers.getRequestHeader("Authorization").get(0).replaceFirst("Basic ", "");
+        if (!sec.basicAuthCheck(encodUsrPwd)) return Response.status(Response.Status.UNAUTHORIZED).entity("User or password is wrong").build();
+
         try {
             Integer stat = employeedao.delete(empnum);
             Logback.logger.info("{}.{}|Try:Deleted!", this.getClass().getSimpleName(), Thread.currentThread().getStackTrace()[1].getMethodName());
